RUBY-686 cleaning up lib structure a bit in prep for auth
* moved: gridfs code into its own folder
* moved: all connection pool related code into its own folder
* moved: all socket and socket utils to connection/socket
* moved: functional modules (read concern, write concern) to functional folder
* fixed: conversion unit test was requiring some specific paths, not needed
