+++ lib/bson/types/object_id.rb
@@ -12,9 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'thread'
-require 'socket'
 require 'digest/md5'
+require 'socket'
 
 module BSON
