+++ lib/mongo.rb
@@ -60,35 +60,16 @@ end
 
 require 'bson'
 
-require 'mongo/util/thread_local_variable_manager'
-require 'mongo/util/conversions'
-require 'mongo/util/support'
-require 'mongo/util/read_preference'
-require 'mongo/util/write_concern'
-require 'mongo/util/core_ext'
-require 'mongo/util/logging'
-require 'mongo/util/node'
-require 'mongo/util/pool'
-require 'mongo/util/pool_manager'
-require 'mongo/util/sharding_pool_manager'
-require 'mongo/util/server_version'
-require 'mongo/util/socket_util'
-require 'mongo/util/ssl_socket'
-require 'mongo/util/tcp_socket'
-require 'mongo/util/unix_socket'
-require 'mongo/util/uri_parser'
-
-
+require 'mongo/utils'
+require 'mongo/exception'
+require 'mongo/functional'
+require 'mongo/connection'
+require 'mongo/collection'
+require 'mongo/cursor'
+require 'mongo/db'
+require 'mongo/gridfs'
 require 'mongo/networking'
 require 'mongo/mongo_client'
 require 'mongo/mongo_replica_set_client'
 require 'mongo/mongo_sharded_client'
 require 'mongo/legacy'
-require 'mongo/collection'
-require 'mongo/cursor'
-require 'mongo/db'
-require 'mongo/exceptions'
-require 'mongo/gridfs/grid_ext'
-require 'mongo/gridfs/grid'
-require 'mongo/gridfs/grid_io'
-require 'mongo/gridfs/grid_file_system'
