+++ lib/mongo/connection/pool.rb
