+++ lib/mongo/connection/pool_manager.rb
