+++ lib/mongo/connection/sharding_pool_manager.rb
