+++ lib/mongo/connection/socket.rb
@@ -1,0 +1,18 @@
+# Copyright (C) 2013 MongoDB, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+require 'mongo/connection/socket/socket_util.rb'
+require 'mongo/connection/socket/ssl_socket.rb'
+require 'mongo/connection/socket/tcp_socket.rb'
+require 'mongo/connection/socket/unix_socket.rb'
