+++ lib/mongo/connection/socket/ssl_socket.rb
@@ -12,9 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'socket'
 require 'openssl'
-require 'timeout'
 
 module Mongo
