+++ lib/mongo/connection/socket/tcp_socket.rb
@@ -12,9 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'socket'
-require 'timeout'
-
 module Mongo
   # Wrapper class for Socket
   #
