+++ lib/mongo/db.rb
@@ -12,9 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'socket'
-require 'thread'
-
 module Mongo
 
   # A MongoDB database.
