+++ lib/mongo/exception.rb
