+++ lib/mongo/functional/logging.rb
