+++ lib/mongo/functional/read_preference.rb
