+++ lib/mongo/functional/uri_parser.rb
