+++ lib/mongo/mongo_client.rb
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'set'
-require 'socket'
 require 'thread'
 
 module Mongo
