+++ lib/mongo/utils/conversions.rb
