+++ lib/mongo/utils/core_ext.rb
