+++ lib/mongo/utils/support.rb
