+++ lib/mongo/utils/thread_local_variable_manager.rb
